+++ lib/igdb_client/query/builder.rb
@@ -10,13 +10,12 @@ module IgdbClient
         @fields    = opts[:fields] || ALL_FIELDS
         @id        = opts[:id]
         @limit     = opts[:limit]
+        @offset    = opts[:offset]
         @search    = opts[:search]
         @sort_by   = opts[:sort_by]
         @sort_direction = opts[:sort_direction]
 
-        raise InvalidArguments, "Cannot combine ID with Search" if @id.present? && @search.present?
-        raise InvalidArguments, "Cannot combine Fields with Exclude" if @fields != "*" && @exclude.present?
-        show_redundant_argument_warning if @id.present? && @limit.present?
+        perform_validations
       end
 
       def build
@@ -29,6 +28,28 @@ module IgdbClient
 
       private
 
+      def perform_validations
+        errors = []
+
+        if @id.present? && @search.present?
+          errors << "Cannot combine ID with Search"
+        end
+
+        if @fields != "*" && @exclude.present?
+          errors << "Cannot combine Fields with Exclude"
+        end
+
+        if @id.present? && @offset.present?
+          errors << "Cannot combine ID with Offset"
+        end
+
+        if errors.any?
+          raise InvalidArguments, errors.join(", ")
+        end
+
+        show_redundant_argument_warning if @id.present? && @limit.present?
+      end
+
       def params
         @params ||= build_params
       end
@@ -39,6 +60,7 @@ module IgdbClient
           fields: Fields::Field.new(@fields),
           id: Fields::Id.new(@id),
           limit: Fields::Limit.new(@limit),
+          offset: Fields::Offset.new(@offset),
           search: Fields::Search.new(@search),
           sort: Fields::Sort.new(@sort_by, @sort_direction)
         }